+++ src/qbits/alia/lamina.clj
@@ -23,12 +23,13 @@
   For other options refer to `qbits.alia/execute` doc"
   ([^Session session query {:keys [success error executor consistency
                                    serial-consistency routing-key
-                                   retry-policy tracing? string-keys? fetch-size
-                                   values timestamp]}]
+                                   retry-policy tracing? string-keys? indepotent?
+                                   fetch-size values timestamp paging-state]}]
      (let [^Statement statement (query->statement query values)]
-       (set-statement-options! statement routing-key retry-policy tracing?
+       (set-statement-options! statement routing-key retry-policy
+                               tracing? indepotent?
                                consistency serial-consistency fetch-size
-                               timestamp)
+                               timestamp paging-state)
        (let [^ResultSetFuture rs-future
              (try
                (.executeAsync session statement)